+++ projects/egeo-elements/src/egeo-library/sds-tooltip/compile.ts
@@ -1,10 +1,10 @@
 import {enableProdMode} from '@angular/core';
 import {platformBrowserDynamic} from '@angular/platform-browser-dynamic';
 
-import { StTooltipElementModule } from './st-tooltip.module';
+import { SdsTooltipElementModule } from './sds-tooltip.module';
 
 enableProdMode();
 
 platformBrowserDynamic()
-    .bootstrapModule(StTooltipElementModule, { ngZone: 'noop'})
+    .bootstrapModule(SdsTooltipElementModule, { ngZone: 'noop'})
     .catch(err => console.error(err));