+++ projects/egeo/src/lib/st-dropdown-menu/st-dropdown-menu-item/st-dropdown-menu-item.component.spec.ts
@@ -9,13 +9,12 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 import { ChangeDetectionStrategy, DebugElement, NO_ERRORS_SCHEMA } from '@angular/core';
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 
 import { StDropDownMenuItem } from '../st-dropdown-menu.interface';
 import { StDropdownMenuItemComponent } from './st-dropdown-menu-item.component';
-import { StBubbleModule } from '../../st-bubble/st-bubble.module';
-import { StBubbleComponent } from '../..';
+import { SdsTooltipComponent } from '../../sds-tooltip';
 
 const item: StDropDownMenuItem = {
    label: 'example 1',
@@ -29,11 +28,10 @@ describe('StDropdownMenuItemComponent', () => {
 
    beforeEach(() => {
       TestBed.configureTestingModule({
-         imports: [StBubbleModule],
          declarations: [StDropdownMenuItemComponent],
          schemas: [NO_ERRORS_SCHEMA]
       })
-         .overrideComponent(StBubbleComponent, {
+         .overrideComponent(SdsTooltipComponent, {
             set: { changeDetection: ChangeDetectionStrategy.Default }
          })
          .compileComponents();  // compile template and css
@@ -255,43 +253,6 @@ describe('StDropdownMenuItemComponent', () => {
 
          expect(fixture.nativeElement.querySelector('.st-dropdown-menu-item i.extra-icon').classList).toContain('icon-info');
          expect(fixture.debugElement.query(By.css('.st-dropdown-menu-item i.extra-icon')).styles.color).toEqual('rgb(45, 78, 92)');
-      });
-
-      describe('If bubble for extra icon is introduced', () => {
-         beforeEach(() => {
-            comp.item = {
-               label: '<b>A</b> M<b>a</b>n <b>a</b>nd <b>a</b> Wom<b>a</b>n',
-               hasHtml: false,
-               value: 'info',
-               extraIconColor: '#fefefe',
-               extraIcon: 'icon-info',
-               extraIconBubble: 'This is an informative text'
-            };
-            fixture.detectChanges();
-         });
-
-         it('it is only displayed when user puts mouse over the icon', (done) => {
-            expect(fixture.nativeElement.querySelector('.st-dropdown-menu-item .extra-icon__bubble i').classList).toContain('icon-info');
-
-            fixture.nativeElement.querySelector('.st-dropdown-menu-item .extra-icon__bubble i').dispatchEvent(new Event('mouseenter'));
-            fixture.detectChanges();
-
-            fixture.whenStable().then(() => {
-               fixture.detectChanges();
-               expect(fixture.nativeElement.querySelector('.st-dropdown-menu-item st-bubble .content').style.visibility).toEqual('visible');
-
-               fixture.nativeElement.querySelector('.st-dropdown-menu-item .extra-icon__bubble i').dispatchEvent(new Event('mouseleave'));
-               fixture.detectChanges();
-
-               fixture.whenStable().then(() => {
-                  fixture.detectChanges();
-
-                  expect(fixture.nativeElement.querySelector('.st-dropdown-menu-item st-bubble .content').style.visibility).toEqual('hidden');
-                  done();
-               });
-
-            });
-         });
       });
    });
 });