+++ projects/egeo/src/lib/st-dropdown-menu/st-dropdown-menu.interface.ts
@@ -31,7 +31,6 @@ export class StDropDownMenuItem {
    selected?: boolean;
    hasHtml?: boolean;
    extraIcon?: string;
-   extraIconBubble?: string;
    extraIconColor?: string;
    disabled?: boolean;
    topSeparator?: boolean;