+++ projects/sds-demo/src/app/demos/sds-checkbox-demo/sds-checkbox-demo.module.ts
@@ -10,7 +10,7 @@
  */
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import {StDemoGeneratorModule, SdsCheckboxModule, StDocsModule, SdsRadioModule, StTooltipModule} from '@stratio/egeo';
+import {StDemoGeneratorModule, SdsCheckboxModule, StDocsModule, SdsRadioModule, SdsTooltipModule} from '@stratio/egeo';
 
 import { SdsCheckboxDemoComponent } from './sds-checkbox-demo.component';
 import { CssPropertyTableModule } from '@app/shared/css-property-table/css-property-table.module';
@@ -19,7 +19,7 @@ import {FormsModule} from '@angular/forms';
 @NgModule({
    imports: [
       CommonModule,
-      StTooltipModule,
+      SdsTooltipModule,
       SdsCheckboxModule,
       SdsRadioModule,
       CssPropertyTableModule,